Moved code to multicommodity folder
