+++ src/rtctools_heat_network/pycml/component_library/heat/multicommodity/__init__.py
@@ -1,0 +1,5 @@
+from .electrolyzer import Electrolyzer
+
+__all__ = [
+    "Electrolyzer",
+]
